+++ nodeconductor/cost_tracking/serializers.py
@@ -43,7 +43,10 @@ class PriceEstimateSerializer(AugmentedSerializerMixin, serializers.HyperlinkedM
         return price_estimate
 
     def get_scope_name(self, obj):
-        return six.text_type(obj.scope or obj.details.get('scope_name'))  # respect to unicode
+        if obj.scope:
+            return six.text_type(obj.scope)
+        if obj.details:
+            return obj.details.get('scope_name')
 
     def get_scope_type(self, obj):
         return ScopeTypeFilterBackend.get_scope_type(obj.content_type.model_class())